% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{acs_dat}
\alias{acs_dat}
\title{Example ACS Small Area Data}
\format{
A data frame with 1,617 rows and 8 variables:
\describe{
\item{MedInc}{Area-level direct estimate of median income.}
\item{MedIncSE}{Standard error of \code{MedInc}.}
\item{SNAPRate}{Area-level SNAP participation rate covariate.}
\item{PovRate}{Area-level poverty rate covariate.}
\item{White}{Area-level proportion identifying as White.}
\item{Black}{Area-level proportion identifying as Black.}
\item{Hispanic}{Area-level proportion identifying as Hispanic.}
\item{Asian}{Area-level proportion identifying as Asian.}
}
}
\usage{
acs_dat
}
\description{
A small example data set for fitting nonlinear Fay-Herriot models. \code{MedInc}
is the direct estimate response and \code{MedIncSE} is its standard error. Use
\code{MedIncSE^2} as the sampling variance when fitting models.
}
\examples{
data(acs_dat)
head(acs_dat)
}
\keyword{datasets}
